Rename modules
We have to rename modules so that the pattern makes sense for external consumers of the package, with `thorium-web_` prefix that scopes classnames
