+++ src/components/Helpers/getReaderClassNames.ts
@@ -1,5 +1,5 @@
 import { ThLayoutUI } from "@/preferences";
-import readerStyles from "../assets/styles/reader.module.css";
+import readerStyles from "../assets/styles/thorium-web.reader.app.module.css";
 import classNames from "classnames";
 
 interface ReaderStyleOptions {
@@ -12,8 +12,8 @@ interface ReaderStyleOptions {
 }
 
 const LAYOUT_CLASSES = {
-  [ThLayoutUI.stacked]: "thoriumWebStackedUI",
-  [ThLayoutUI.layered]: "thoriumWebLayeredUI",
+  [ThLayoutUI.stacked]: "thorium_web_stackedUI",
+  [ThLayoutUI.layered]: "thorium_web_layeredUI",
 } as const;
 
 export function getReaderClassNames(options: ReaderStyleOptions): string {
@@ -28,11 +28,11 @@ export function getReaderClassNames(options: ReaderStyleOptions): string {
 
   return classNames(
     readerStyles.thoriumWebReaderShell,
-    isScroll ? "thoriumWebIsScroll" : "thoriumWebIsPaged",
-    isImmersive && "thoriumWebIsImmersive",
-    isHovering && "thoriumWebIsHovering",
-    isFXL ? "thoriumWebIsFXL" : "thoriumWebIsReflow",
+    isScroll ? "thorium_web_isScroll" : "thorium_web_isPaged",
+    isImmersive && "thorium_web_isImmersive",
+    isHovering && "thorium_web_isHovering",
+    isFXL ? "thorium_web_isFXL" : "thorium_web_isReflow",
     LAYOUT_CLASSES[layoutUI],
-    breakpoint ? `thoriumWebIs${ breakpoint.charAt(0).toUpperCase() + breakpoint.slice(1) }` : undefined
+    breakpoint ? `thorium_web_is${ breakpoint.charAt(0).toUpperCase() + breakpoint.slice(1) }` : undefined
   );
 }
