+++ src/components/Settings/StatefulNumberField.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import readerSharedUI from "../assets/styles/readerSharedUI.module.css";
-import settingsStyles from "./assets/styles/settings.module.css";
+import readerSharedUI from "../assets/styles/thorium-web.button.module.css";
+import settingsStyles from "./assets/styles/thorium-web.reader.settings.module.css";
 
 import { ThNumberField, ThNumberFieldProps } from "@/core/Components/Settings/ThNumberField";
