+++ src/components/Sheets/StatefulDockedSheet.tsx
@@ -5,8 +5,8 @@ import React, { useCallback, useRef } from "react";
 import { StatefulSheet } from "./models/sheets";
 import { ThDockingKeys, ThSheetHeaderVariant, ThLayoutDirection } from "@/preferences/models/enums";
 
-import sheetStyles from "./assets/styles/sheets.module.css";
-import readerSharedUI from "../assets/styles/readerSharedUI.module.css";
+import sheetStyles from "./assets/styles/thorium-web.sheets.module.css";
+import readerSharedUI from "../assets/styles/thorium-web.button.module.css";
 
 import { ThDockedPanel } from "@/core/Components/Containers/ThDockedPanel";
 import { ThContainerHeader } from "@/core/Components/Containers/ThContainerHeader";
